+++ Engine/__init__.py
@@ -6,7 +6,7 @@ from .Kernel.Components.control import Key, MouseButton, mEvent, kEvent, Keys
 from .Kernel.kernel import log_system, setDebug
 from .Kernel.fonts import fonts
 from .Kernel.icons import icons
-from .Kernel.math import Math
+from .Kernel.math import Math, Random
 
 # CONTROL