+++ Engine/Physics/config.py
@@ -1,0 +1,8 @@
+gravity = 16.4
+
+def setGravity(g:float | int = 16.4):
+    global gravity
+    gravity = g
+
+def getGravity():
+    return gravity